london/requirements.txt
numpy
geopy
pandas
seaborn
networkx
matplotlib
torch_geometric
torch --index-url https://download.pytorch.org/whl/cpu
mapie
torchcp
torch-cluster
openpyxl
torch-geometric-temporal
plotly[express]
nbformat>=4.2.0
dash
dash-bootstrap-components